cloud-functions/consultar_telefono/requirements.txt
google-cloud-spanner
functions-framework
